Split renderWithProviders and renderRouterWithProviders in 2 files
