+++ src/testUtil/renderWithProviders.tsx
@@ -2,9 +2,7 @@ import { PreloadedState } from "@reduxjs/toolkit";
 import { render } from "@testing-library/react";
 import { PropsWithChildren } from "react";
 import { Provider } from "react-redux";
-import { RouterProvider } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
-import { getComponentRouter, router } from "../routers/routers";
 import { RootState, setupStore, store } from "../store/store";
 import GlobalStyles from "../styles/GlobalStyles";
 import { theme } from "../styles/theme";
@@ -27,16 +25,4 @@ export const renderWithProviders = (
   };
 
   return render(ui, { wrapper: Wrapper });
-};
-
-export const renderRouterWithProviders = (
-  ui?: React.ReactElement,
-  preloadedState?: PreloadedState<RootState>
-) => {
-  const routerWithProvider = ui ? getComponentRouter(ui) : router;
-
-  return renderWithProviders(
-    <RouterProvider router={routerWithProvider}></RouterProvider>,
-    preloadedState
-  );
 };
